token-2022: Bump to new major version 5.0.0
#### Problem

Token-2022 has had breaking changes, but it's still on major version 4,
which is causing the downstream jobs in the Agave repo to fail when
trying to choose the right version of token-2022.

#### Solution

Make the job a bit easier by bumping the major version, removing
ambiguity for the resolver during downstream jobs.